+++ CHANGES.md
@@ -16,6 +16,8 @@ This file is a best-effort approach to solving this issue; we will do our best b
 
 ### New API
 
+* (network) Added a function to detect IPv4 APIPA addresses (169.254.0.0/16).
+
 ### Changes to existing API
 
 * (internet-apps) Added a parameter to the RADVD helper to announce a prefix without the autoconfiguration flag.